TP-44b update: generator for 1 to menu
