+++ test/newdata/persons.json
@@ -16,5 +16,22 @@
       "сценарист",
       "режиссер"
     ]
+  },
+  {
+    "name": "Шон Коннери",
+    "original_name": "Sean Connery",
+    "birthday": "1930.08.25",
+    "death": "2020.10.31",
+    "growth": 1.9,
+    "genres": [
+      "драма",
+      "боевик",
+      "триллер"
+    ],
+    "Professions": [
+      "актер",
+      "продюсер",
+      "режиссер"
+    ]
   }
 ]
